deleteUser function is done

+++ src/main/java/com/codecool/smartest/jwt/JwtUserDetails.java
@@ -14,7 +14,7 @@ public class JwtUserDetails implements UserDetails {
 
     private static final long serialVersionUID = 5155720064139820502L;
 
-    private final Long id;
+    private final int id;
     private final String username;
     private final String password;
     private final Collection<? extends GrantedAuthority> authorities;
@@ -33,7 +33,7 @@ public class JwtUserDetails implements UserDetails {
     }
 
     @JsonIgnore
-    public Long getId() {
+    public int getId() {
         return id;
     }
